Update contributor list. Remove unnecessary casts.

+++ Source/Urho3D/Math/Color.cpp
@@ -66,10 +66,10 @@ Vector3 Color::ToHSV() const
 
 void Color::FromUInt(unsigned color)
 {
-    a_ = (float)(((color >> 24) & 0xff) / 255.0f);
-    b_ = (float)(((color >> 16) & 0xff) / 255.0f);
-    g_ = (float)(((color >> 8)  & 0xff) / 255.0f);
-    r_ = (float)(((color >> 0)  & 0xff) / 255.0f);
+    a_ = ((color >> 24) & 0xff) / 255.0f;
+    b_ = ((color >> 16) & 0xff) / 255.0f;
+    g_ = ((color >> 8)  & 0xff) / 255.0f;
+    r_ = ((color >> 0)  & 0xff) / 255.0f;
 }
 
 void Color::FromHSL(float h, float s, float l, float a)
